refactor: staticMesh 위치 이동.
