+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Mesh/StaticMesh.h
@@ -4,6 +4,7 @@
 #include "Components/Material/Material.h"
 #include "Define.h"
 
+struct FStaticMesh;
 
 class UStaticMesh : public UObject
 {
@@ -21,10 +22,7 @@ public:
     FStaticMesh* GetRenderData() const { return staticMeshRenderData; }
 
     //ObjectName은 경로까지 포함
-    FWString GetOjbectName() const
-    {
-        return staticMeshRenderData->ObjectName;
-    }
+    FWString GetOjbectName() const;
 
     void SetData(FStaticMesh* renderData);
