+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Engine/FObjLoader.cpp
@@ -4,6 +4,8 @@
 #include "Components/Material/Material.h"
 #include "Components/Mesh/StaticMesh.h"
 
+#include "Asset/StaticMeshAsset.h"
+
 #include <fstream>
 #include <sstream>
